small refactor and formatting
