+++ src/marine-licences/api/controllers/submit-marine-licence.test.js
@@ -211,7 +211,9 @@ describe('POST /marine-licence/submit', () => {
         mockHandler
       )
 
-      expect(createTaskList).toHaveBeenCalledWith(mockMarineLicence, false)
+      expect(createTaskList).toHaveBeenCalledWith(mockMarineLicence, false, {
+        marinePlanPolicyResponseCount: 0
+      })
     })
   })
