+++ src/marine-licences/api/helpers/createTaskList.js
@@ -8,7 +8,6 @@ import {
   getStatusFromRequiredFields
 } from '../../../shared/helpers/task-list-utils.js'
 import { MARINE_PLAN_POLICY_JOB_STATUS } from '../../constants/marine-licence.js'
-import { filterCurrentPolicyResponses } from './marine-plan-policies/filter-policy-responses.js'
 
 const ACTIVITY_DETAILS_FIELDS = [
   'activities',
@@ -214,7 +213,7 @@ const getFeeEstimateStatus = (feeEstimate) => {
   return COMPLETED
 }
 
-const getMarinePlanPolicyStatus = (marineLicence) => {
+const getMarinePlanPolicyStatus = (marineLicence, completed) => {
   if (
     marineLicence.marinePlanPolicyJob !== MARINE_PLAN_POLICY_JOB_STATUS.READY
   ) {
@@ -222,10 +221,6 @@ const getMarinePlanPolicyStatus = (marineLicence) => {
   }
 
   const total = marineLicence.marinePlanPoliciesCount ?? 0
-  const { count: completed } = filterCurrentPolicyResponses(
-    marineLicence.marinePlanPolicies,
-    marineLicence.marinePlanPolicyResponses
-  )
 
   if (completed >= total) {
     return COMPLETED
@@ -238,7 +233,11 @@ const getMarinePlanPolicyStatus = (marineLicence) => {
   return IN_PROGRESS
 }
 
-export const createTaskList = (marineLicence, isCitizen = false) => {
+export const createTaskList = (
+  marineLicence,
+  isCitizen = false,
+  { marinePlanPolicyResponseCount = 0 } = {}
+) => {
   const tasks = {
     projectName: (value) => (value ? COMPLETED : INCOMPLETE),
     ...(!isCitizen && {
@@ -254,7 +253,8 @@ export const createTaskList = (marineLicence, isCitizen = false) => {
     publicConsultation: (value) => (value ? COMPLETED : INCOMPLETE),
     publicRegister: (value) => (value ? COMPLETED : INCOMPLETE),
     waterFrameworkDirective: (value) => getWaterFrameworkDirectiveStatus(value),
-    marinePlanPolicies: () => getMarinePlanPolicyStatus(marineLicence)
+    marinePlanPolicies: () =>
+      getMarinePlanPolicyStatus(marineLicence, marinePlanPolicyResponseCount)
   }
 
   return buildTaskList(marineLicence, tasks)
